Resolved issue with the $HOME environment not set.
The HOME variable was not set when the Puppet Master executes the
node_terminus. Therefore it is necessary to avoid relying on the HOME variable.

Added new methods for creating and deleting SimpleDB domains. This adds
convenience for users that do not want to use an API to create the domain
before using the Puppet SDB ENC.